requirements.txt
farm-ng-amiga
fastapi
uvicorn[standard]
